brainbox/deciders/voice_generation/piper_training/requirements.lock.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.129.0
flask==3.1.2
h11==0.16.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpickle==4.1.1
markupsafe==3.0.3
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
websocket-client==1.9.0
websockets==16.0
werkzeug==3.1.5

aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
cffi==1.17.1
coloredlogs==15.0.1
Cython==0.29.37
decorator==5.1.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
humanfriendly==10.0
idna==3.10
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.12.0
llvmlite==0.43.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
numba==0.60.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
packaging==24.2
piper-phonemize==1.1.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.3
pycparser==2.22
pytorch-lightning==1.9.5
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.13.1
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.3
threadpoolctl==3.5.0
torchmetrics==1.5.2
tqdm==4.67.1
yarl==1.18.3